requirements.txt
strands-agents
strands-agents-tools
pytest
